pyproject.toml
[project]
name = "core-engine"
version = "0.1.0"
description = "Core engine using langgraph"
requires-python = ">=3.11"

dependencies = [
    "langchain>=0.3.27",
    "langgraph>=0.6.10",
    "langchain-anthropic>=0.3.0",
    "pydantic>=2.12.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "motor>=3.6.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-env>=1.1.0",
    "httpx>=0.27.0",
    "ruff>=0.14.0",
    "mypy>=1.0.0",
    "testcontainers[mongodb]>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning:testcontainers.core.waiting_utils",
    "ignore::DeprecationWarning:testcontainers.mongodb",
]
env = [
    "PYTHONPYCACHEPREFIX=.cache/python",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "B008",  # Allow function calls in argument defaults (FastAPI Depends pattern)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_reexport = true
ignore_missing_imports = true